Split out Qt code for volume viewer
