+++ glue_vispy_viewers/tests/test_session_back_compat.py
@@ -9,7 +9,7 @@ from numpy.testing import assert_equal
 
 from glue.core.state import GlueUnSerializer
 
-from ..volume.volume_viewer import QMessageBox
+from qtpy.QtWidgets import QMessageBox
 
 DATA = os.path.join(os.path.dirname(__file__), 'data')
