+++ glue_vispy_viewers/volume/qt/layer_style_widget.ui
