+++ glue_vispy_viewers/volume/volume_viewer.py
@@ -3,68 +3,29 @@ import numpy as np
 
 from glue.config import settings
 
-from qtpy.QtWidgets import QMessageBox
-from qtpy.QtCore import QTimer
-
-from glue.core.data import BaseData
-from glue.core.link_helpers import LinkSame
-
-from ..common.qt.data_viewer import BaseVispyViewer
+from ..common.vispy_data_viewer import BaseVispyViewerMixin
 from .layer_artist import VolumeLayerArtist
-from .layer_style_widget import VolumeLayerStyleWidget
 from .viewer_state import Vispy3DVolumeViewerState
-from .layer_state import VolumeLayerState
-from .volume_visual import MultiVolume
 
 from ..scatter.layer_artist import ScatterLayerArtist
-from ..scatter.qt.layer_style_widget import ScatterLayerStyleWidget
+from .volume_visual import MultiVolume
 
 from ..common import tools as _tools, selection_tools  # noqa
 from . import volume_toolbar  # noqa
 
 
-class VispyVolumeViewer(BaseVispyViewer):
+class VispyVolumeViewerMixin(BaseVispyViewerMixin):
 
     LABEL = "3D Volume Rendering"
 
     _state_cls = Vispy3DVolumeViewerState
-    _layer_style_widget_cls = {VolumeLayerArtist: VolumeLayerStyleWidget,
-                               ScatterLayerArtist: ScatterLayerStyleWidget}
 
-    tools = BaseVispyViewer.tools + ['vispy:lasso', 'vispy:rectangle',
-                                     'vispy:circle', 'volume3d:floodfill']
+    tools = BaseVispyViewerMixin.tools + ['vispy:lasso', 'vispy:rectangle',
+                                          'vispy:circle', 'volume3d:floodfill']
 
-    def __init__(self, *args, **kwargs):
+    def setup_widget_and_callbacks(self):
 
-        super(VispyVolumeViewer, self).__init__(*args, **kwargs)
-
-        # We now make it so that is the user clicks to drag or uses the
-        # mouse wheel (or scroll on a trackpad), we downsample the volume
-        # rendering temporarily.
-
-        canvas = self._vispy_widget.canvas
-
-        canvas.events.mouse_press.connect(self.mouse_press)
-        canvas.events.mouse_wheel.connect(self.mouse_wheel)
-        canvas.events.mouse_release.connect(self.mouse_release)
-
-        viewbox = self._vispy_widget.view.camera.viewbox
-
-        viewbox.events.mouse_wheel.connect(self.camera_mouse_wheel)
-        viewbox.events.mouse_move.connect(self.camera_mouse_move)
-        viewbox.events.mouse_press.connect(self.camera_mouse_press)
-        viewbox.events.mouse_release.connect(self.camera_mouse_release)
-
-        self._downsampled = False
-
-        # For the mouse wheel, we receive discrete events so we need to have
-        # a buffer (for now 250ms) before which we consider the mouse wheel
-        # event to have stopped.
-
-        self._downsample_timer = QTimer()
-        self._downsample_timer.setInterval(250)
-        self._downsample_timer.setSingleShot(True)
-        self._downsample_timer.timeout.connect(self.mouse_release)
+        super().setup_widget_and_callbacks()
 
         # We need to use MultiVolume instance to store volumes, but we should
         # only have one per canvas. Therefore, we store the MultiVolume
@@ -73,7 +34,7 @@ class VispyVolumeViewer(BaseVispyViewer):
         # Set whether we are emulating a 3D texture. This needs to be
         # enabled as a workaround on Windows otherwise VisPy crashes.
         emulate_texture = (sys.platform == 'win32' and
-                           sys.version_info[0] < 3)
+                            sys.version_info[0] < 3)
 
         multivol = MultiVolume(emulate_texture=emulate_texture,
                                bgcolor=settings.BACKGROUND_COLOR)
@@ -84,27 +45,6 @@ class VispyVolumeViewer(BaseVispyViewer):
         self.state.add_callback('resolution', self._update_resolution)
         self._update_resolution()
 
-        # We do this here in addition to in the volume viewer itself as for
-        # some situations e.g. reloading from session files, a clip_data event
-        # isn't emitted.
-        # FIXME: needs to be done after first layer added
-        # self._update_clip(force=True)
-
-    def mouse_press(self, event=None):
-        if self.state.downsample:
-            if hasattr(self._vispy_widget, '_multivol') and not self._downsampled:
-                self._vispy_widget._multivol.downsample()
-                self._downsampled = True
-
-    def mouse_release(self, event=None):
-        if self.state.downsample:
-            if hasattr(self._vispy_widget, '_multivol') and self._downsampled:
-                self._vispy_widget._multivol.upsample()
-                self._downsampled = False
-                self._vispy_widget.canvas.render()
-
-        self._update_slice_transform()
-        self._update_clip()
 
     def _update_clip(self, force=False):
         if hasattr(self._vispy_widget, '_multivol'):
@@ -129,19 +69,6 @@ class VispyVolumeViewer(BaseVispyViewer):
         self._update_slice_transform()
         self._update_clip()
 
-    def mouse_wheel(self, event=None):
-        if self.state.downsample:
-            if hasattr(self._vispy_widget, '_multivol'):
-                if not self._downsampled:
-                    self.mouse_press()
-        self._downsample_timer.start()
-        if event is not None:
-            event.handled = True
-
-    def resizeEvent(self, event=None):
-        self.mouse_wheel()
-        super(VispyVolumeViewer, self).resizeEvent(event)
-
     def get_data_layer_artist(self, layer=None, layer_state=None):
         if layer.ndim == 1:
             cls = ScatterLayerArtist
@@ -162,23 +89,17 @@ class VispyVolumeViewer(BaseVispyViewer):
 
         if data.ndim == 1:
             if first_layer_artist:
-                QMessageBox.critical(self, "Error",
-                                     "Can only add a scatter plot overlay once "
-                                     "a volume is present",
-                                     buttons=QMessageBox.Ok)
-                return False
+                raise Exception("Can only add a scatter plot overlay once "
+                                "a volume is present")
         elif data.ndim == 3:
             if not self._has_free_volume_layers:
                 self._warn_no_free_volume_layers()
                 return False
         else:
-            QMessageBox.critical(self, "Error",
-                                 "Data should be 1- or 3-dimensional ({0} dimensions "
-                                 "found)".format(data.ndim),
-                                 buttons=QMessageBox.Ok)
-            return False
+            raise Exception("Data should be 1- or 3-dimensional ({0} dimensions "
+                            "found)".format(data.ndim))
 
-        added = super(VispyVolumeViewer, self).add_data(data)
+        added = super().add_data(data)
 
         if added:
 
@@ -204,7 +125,7 @@ class VispyVolumeViewer(BaseVispyViewer):
             self._warn_no_free_volume_layers()
             return False
 
-        added = super(VispyVolumeViewer, self).add_subset(subset)
+        added = super().add_subset(subset)
 
         if added:
             self._show_free_layer_warning = True
@@ -218,17 +139,12 @@ class VispyVolumeViewer(BaseVispyViewer):
 
     def _warn_no_free_volume_layers(self):
         if getattr(self, '_show_free_layer_warning', True):
-            QMessageBox.critical(self, "Error",
-                                 "The volume viewer has reached the maximum number "
-                                 "of volume layers. To show more volume layers, remove "
-                                 "existing layers and try again. This error will not "
-                                 "be shown again unless the limit is reached again in "
-                                 "the future.",
-                                 buttons=QMessageBox.Ok)
-            self._show_free_layer_warning = False
+            raise Exception("The volume viewer has reached the maximum number "
+                            "of volume layers. To show more volume layers, remove "
+                            "existing layers and try again.")
 
     def _update_appearance_from_settings(self, message):
-        super(VispyVolumeViewer, self)._update_appearance_from_settings(message)
+        super()._update_appearance_from_settings(message)
         if hasattr(self._vispy_widget, '_multivol'):
             self._vispy_widget._multivol.set_background(settings.BACKGROUND_COLOR)
 
@@ -236,47 +152,7 @@ class VispyVolumeViewer(BaseVispyViewer):
         if hasattr(self._vispy_widget, '_multivol'):
             self._update_clip()
 
-    @classmethod
-    def __setgluestate__(cls, rec, context):
-
-        viewer = super(VispyVolumeViewer, cls).__setgluestate__(rec, context)
-
-        if rec.get('_protocol', 0) < 2:
-
-            # Find all data objects in layers (not subsets)
-            layer_data = [layer.layer for layer in viewer.state.layers
-                          if (isinstance(layer, VolumeLayerState) and
-                              isinstance(layer.layer, BaseData))]
-
-            if len(layer_data) > 1:
-                reference = layer_data[0]
-                for data in layer_data[1:]:
-                    if data not in reference.pixel_aligned_data:
-                        break
-                else:
-                    return viewer
-
-            buttons = QMessageBox.Yes | QMessageBox.No
-            message = ("The 3D volume rendering viewer now requires datasets to "
-                       "be linked in order to be shown at the same time. Are you "
-                       "happy for glue to automatically link your datasets by "
-                       "pixel coordinates?")
-
-            answer = QMessageBox.question(None, "Link data?", message,
-                                          buttons=buttons,
-                                          defaultButton=QMessageBox.Yes)
-
-            if answer == QMessageBox.Yes:
-                for data in layer_data[1:]:
-                    if data not in reference.pixel_aligned_data:
-                        for i in range(3):
-                            link = LinkSame(reference.pixel_component_ids[i],
-                                            data.pixel_component_ids[i])
-                            viewer.session.data_collection.add_link(link)
-
-        return viewer
-
     def __gluestate__(self, context):
-        state = super(VispyVolumeViewer, self).__gluestate__(context)
+        state = super().__gluestate__(context)
         state['_protocol'] = 2
         return state
