+++ setup.cfg
@@ -45,7 +45,7 @@ all =
 glue_vispy_viewers.common = *.png
 glue_vispy_viewers.common.qt = *.ui
 glue_vispy_viewers.scatter.qt = *.ui
-glue_vispy_viewers.volume = *.ui
+glue_vispy_viewers.volume.qt = *.ui
 glue_vispy_viewers.tests = data/*.glu
 glue_vispy_viewers.extern = README.md, VISPY_LICENSE
 glue_vispy_viewers.extern.vispy = io/_data/*.npy, html/static/js/*.js, app/tests/*.ui
